Cubari release_date fix + volume null fix (#8163)
* Fix for Cubari API response not containing chapter release_date

Cubari API responses don't contain release_date information when the Cubari source is Imgur.
The minor change made with this commit makes the existence of the release_date json property for the parsing process optional.

* Fix for Cubari volume "null" and cleaned up chapter name generation

If the volume is set to null in the json, it's now recognized as not specified.
Also cleaned up the chapter name concatenation.