Add extension: Real Life Comics (#8653)
* add icon for reallifecomics source

* add required manifest and gradle file

* add reallifecomics source

original webcomic link - https://reallifecomics.com

* page index starts at 0

* initialise Page correctly

imageUrl was what I was trying to set

* parse date_upload for chapterList differently

posts before 2014 don't have the date in the link. However, they are
still arranged in a Calendar. Use this calendar to find a date.

* skip years with no archive pages

* cleanup and edits

* now able to parse all years without link issues

* add an extra space to chapter names

for aesthetics! also follows what is shown on the website

* have a simple Search feature

since a search just throwing an error seemed weird