Fetch language explicit chapters links for CommitStrip (#8922)
* fix Ext icon in Extensions list

replaces android-studio-created image assets that lead to blank images in
Extensions list (in app and in https://tachiyomi.org/extensions as
well). Usually due to there being fg bg and xml files and not just a
single png file

Co-authored-by: as280093 as280093@gmail.com

* fix chapters not parsed for single page archive

latest archives (as of writing, 2021) can just have one page in which
case parsing to find total no of pages doesn't really work. In these
cases, just send a single request and that'd be all

* use lang specific chapter links

be default each chapter gives a link like this:
 https://www.commitstrip.com/2020/03/23/a-mask-for-everyone/
which then is redirected to the lang by remembering browser state
instead we can force the link to be the lang specific one i.e.
 https://www.commitstrip.com/{en,fr}/2020/03/23/a-mask-for-everyone/

* fix link to WebView

sets the correct to be opened when using Open in WebView or Share

* Revert "fix chapters not parsed for single page archive"

This reverts commit 4722790a8db905ba18259758cbf9ed47890e9eff.
Didn't really work like I expected. Time to try another route

* bump commitstrip.extversioncode