Update DragonBallMultiverse to new site layout (#9006)
* update link to manga

Closes https://github.com/tachiyomiorg/tachiyomi-extensions/issues/8568

* bump dbm.extvercode

* add more titles to ext

* add fetchMangaDetails override

so that refreshing manga works

* update chapterListSelector

so that it works on all titles, old and new

* refactor parsing chapter lists

since the selector was changed

* remember chapter currentTimeMillis

kind of a hack for setting `chapter.date_upload` cleverly so that it
doesn't update itself during library update
refer tachiyomiorg@50859e7

* refactor chapterFromElement again

because name inside SChapter.create().apply{} was ambiguous. made it
explicit to `chapter.name`

* fix value set to `chapter.chapter_number`

* revert chapterListParse

remembering currentTimeInMillis doesn't really work for now. Can just
add it if requested later

This commit reverts 459057829

* remove unused add override for chapterFromElement()

* create absolute links instead of fetching them

so that I can set the chapter names correctly