FMReader: WeLoveManga: Fix page list parsing (#9196)
* FMReader: WeLoveManga: Fix page list parsing

They appear to have changed how their page list works, to instead be driven
off an attribute with a seemingly randomly generated name.  That name is
then emitted as part of the page script.  We can just rely on that
attribute being very long instead.

Fixes #9120

* FMReader: WeLoveManga: Fix imports.

They were not copied over correctly when copying the (edited) generated
code back to the source.