[Komga] Fix link to WebView/Share (#9361)
* use the correct link for viewing in browser

api url on browser just provides a json. For viewing in browser,
we need the normal url instead

* just use `substringAfter`

was trying out another way using `substringAfterLast` but dropped that.
Now `substringAfterLast` isn't required and `substringAfter` suffices

* make fetchMangaDetails not use mangaDetailsRequest

internally, `fetchMangaDetails` uses `mangaDetailsRequest` to create a
new call. However, since `mangaDetailsRequest` is now fixed to use a
normal browser URL rather than the API URL, it no longer works for
`fetchMangaDetails`. So, it needs to be manually overriden to make its
own call

* use hardcoded ID values instead of computing

this was done because of the following:
 - IDs are same as before
 - no need to rely on calculation to be correct everytime
 - avoids unnecessary copy of base class code
 - sets it in the constructor itself

* increment komga.extVersionCode

also update CHANGELOG with list of changes

* linting and remove unused imports

* Revert "linting and remove unused imports"

This reverts commit e13396badef53fced2700266182bbf53a67124cf.

* Revert "use hardcoded ID values instead of computing"

This reverts commit 44f063904d5969fb6d90c9ffad5c33aff50b995f.

* refactor `mangaDetailsRequest`

Do it in a simpler manner without having to create another variable

Co-authored-by: Gauthier <gotson@users.noreply.github.com>

Co-authored-by: Gauthier <gotson@users.noreply.github.com>