latisbooks: add text pages for accompanying text (#11102)
Not all pages have accompanying text, but it's nice to be able to see
the text when there is some. Examples are `No Comic (03/14/22)` and
`Meidri (Interspecies Reviewers)`
This is done in a nearly identical way to the xkcd source's alt-text
pages, with nearly identical code.
Also handle the extremely rare case where there are multiple images for
the same page. The only instance I can recall is `Page 23+24`