Manhwa latino (#13054)
* First commit Mahnwa-Latino Extension.

* manhwa-latino: Implementing find function

TODO: only works with text, filter not implemented yet

* manhwa-latino Add Parser to finde the information of Manhwa-Latino

The parser 'ManhwaLatinoSiteParser.kt' make the whole magic to find the
information of the Website.

Mudularize The code and adding documentation

* manhwa-latino: Adding Logos der Extension

TODO: I am not to happy, i will check it later.

* manhwa-latino: Adding Tags to Genre Combobox

* manhwa-latino: Adding Headers to prevent error 404

The headers are necesary to prevent error 403 by downloading images.

* manhwa-latino: Tags addded into Manga Description Page

Status from Manga readed from Tags

* manhwa-latino: Modularize Code

* manhwa-latino: Adding Uploaddate for Chapters

* manhwa-latino: Bug to get Chapter Number fixed

* manhwa-latino: Logo 0.2

* manhwa-latino: Versionb 1.2.10

Adding Comments to ManhwaLatinoSiteParser

* manhwa-latino: Remove logo_model directory

* manhwa-latino: Show Seconds after Release a new Chapter

* manhwa-latino: Fix for Cloudflare

* manhwa-latino: Try to fix Cloudflare ByPass

Adding isNsfw to build.gradle

* manhwa-latino: Update from parser after website update

* manhwa-latino: Deactivate lastest manga and new Client

The new website from Manhwa-Latino doesnt show lastest mangas.

* manhwa-latino: Adding Manifest to load url-adress

The extesion will bi able to open and find mangas with the url.
TODO: first version. NOT READY YET.

* manhwa-latino: Fix to find Manga Status information

* Manhwa-Latino: The Extension can open a Link

Now the extension can recognize a link from Manhwa-Latino and
show the manga from the link in the app.

* Manhwa-Latino: Modularize Code and Activity for Url added

* Manhwa-Latino: Add Constant-File.

This file contain all contants for this extension.

* manhwa-latino Url from the Website has changed.

from https://manhwa-latino.com => https://manhwa-es.com

* manhwa-latino Url update on manifest

* manhwa-latino Fix to get ThumbnailUrl and Url from Slider

The Slider from Manhwa-Latino was updated and the parser was updated as
well.

* manhwa-latino: Using Cloudflare Client to connect.

Manhwa-Latino now Manhwa-Es are using cloudflare right now.
This workaround fix make possible the connection from the app to the
Website.

* Update the address to every image

* Update Manhwa-Latino Address

* Updating popularMangaNextPageSelector

* Remove CLoudFlare Client

* Manhwa-Latino: Update Image Src-Tag

The website has changed the attribute where the adress to the images
are.

* Mahnwa-Latino: Remove client, because it is not overwritten

Original Client is:
	override val client: OkHttpClient = network.client.newBuilder().build()

it is not overwritten, that is why it can remove.

* manhwa-latino: Remove Comments

Co-authored-by: Luis Beroiza <luisalberto.beroizaosses@intern.osp-dd.de>
Co-authored-by: AlberyKous <extreme_rvd@hotmail.com>