New Lib: Text Interceptor (#13859)
* move TextInterceptor to `lib`

to be used for author notes, image alt texts and what not

* refactor Tapastic to use `lib:extensions-lib`

* Refactor Webtoons to use `:lib-textinterceptor`

this turned out to be more complicated than I thought it'd be
TextInterceptor was used for Author Notes which, looking at previous messages,
was only intended to be added to Webtoons Source and not the entire webtoons
multisrc (i.e. WebtoonsTranslate and DongmanManhua don't seem to be making use
of the Show Author's Notes setting). This was in my favor, since having
to deal with additional.gradle to add dependencies to multisrc files doesn't
seem to work... I'll ask previous contributors just in case

* Fix `json` access

missed this while copying over code from `generated-src` to `multisrc/{overrides,src}`

* remove unused import

* make HOST name more clear

couldve used a better schema but that's something for some other time
also put the HOST in the lib itself so that one doesn't lose track of it in the extensions

* use android provided methods instead of hardcoding

based on https://github.com/tachiyomiorg/tachiyomi-extensions/pull/13859/files#r996276738
that suggested the following SO answer: https://github.com/tachiyomiorg/tachiyomi-extensions/pull/13859/files

* remove unused import

* move url generation to helper function

* fix error

oops sorry for that
happened when I was copy pasting back and forth between two sources and one generated source

Co-authored-by: Navaneeth M Nambiar <nmnambiar@hornbill>