latisbooks: fix page parsing, fix cover art (#17135)
* latisbooks: fix page parsing

Likely from a preprocessor change on the site's end, the element
containing the images had changed.

* latisbooks: fix cover art

The cover art being used for the comic is pulled from the "latest page"
preview on the archive page, which is a square image that tends to not
represent the comic very well. Additionally, recently it has failed to
load properly for whatever reason. As a remedy, I chose page 198 to be 
the proper cover. I chose it as it is an interstitial cover page
featuring the main character, with no text or elements that tie it to a
specific arc.

* latisbooks: version bump